server/requirements/requirements_model.txt
celery==5.5.3
httpx==0.28.1
monai==1.5.1
numpy==2.2.6
opencv_python==4.12.0.88
Pillow==12.0.0
python-dotenv==1.2.1
segmentation_models_pytorch==0.5.0
torch==2.9.0
torchvision==0.24.0
